requirements.txt
opencv-python
ultralytics
torch
torchvision
numpy
Pillow